+++ eth/tracers/native/call_flat.go
@@ -144,10 +144,11 @@ func newFlatCallTracer(ctx *tracers.Context, cfg json.RawMessage, chainConfig *p
 	ft := &flatCallTracer{tracer: t, ctx: ctx, config: config, chainConfig: chainConfig}
 	return &tracers.Tracer{
 		Hooks: &tracing.Hooks{
-			OnTxStart: ft.OnTxStart,
-			OnTxEnd:   ft.OnTxEnd,
-			OnEnter:   ft.OnEnter,
-			OnExit:    ft.OnExit,
+			OnTxStart:                 ft.OnTxStart,
+			OnTxEnd:                   ft.OnTxEnd,
+			OnEnter:                   ft.OnEnter,
+			OnExit:                    ft.OnExit,
+			OnSystemTxFixIntrinsicGas: ft.OnSystemTxFixIntrinsicGas,
 		},
 		Stop:      ft.Stop,
 		GetResult: ft.GetResult,